% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getFileName}
\alias{getFileName}
\title{Get file name from complete path}
\usage{
getFileName(filePath = NULL)
}
\arguments{
\item{filePath}{complete path to a file}
}
\value{
file name and its extension
}
\description{
Get file name from complete path
}
\examples{
file <- system.file(
    "extdata", "query.fa", package = "phylosophy", mustWork = TRUE
)
getFileName(file)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
