% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getSeqID}
\alias{getSeqID}
\title{Get sequence IDs from fasta file}
\usage{
getSeqID(file = NULL)
}
\arguments{
\item{file}{fasta file}
}
\value{
list object contains sequence IDs from input file
}
\description{
Get sequence IDs from fasta file
}
\examples{
file <- system.file(
    "extdata", "query.fa", package = "phylosophy", mustWork = TRUE
)
getSeqID(file)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
