% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{randFn}
\alias{randFn}
\title{Create random strings}
\usage{
randFn(n = 5000)
}
\arguments{
\item{n}{number of output strings}
}
\value{
n strings created by 10 random characters
}
\description{
Create random strings
}
\examples{
randFn(1)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
