% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPhylosophy.R
\name{runPhylosophy}
\alias{runPhylosophy}
\title{Run phylosophy app}
\usage{
runPhylosophy()
}
\value{
A shiny application - GUI version of phylosophy
}
\description{
Run phylosophy app
}
\examples{
?runPhylosophy
\dontrun{
runPhylosphy()
}
}
